# server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Quiz Application API
  version: 1.0.0
  description: API documentation for Quiz Application
servers:
  - url: http://localhost:4000/api
    description: Development server
paths:
...
  /quiz/comments:
    post:
      tags:
        - Quiz
      summary: Add comment to quiz
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
      responses:
        '201':
          description: Comment added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment added"
                  payload:
                    type: object
                    properties:
                      comment:
                        type: object
        '400':
          description: Comment text cannot be empty
        '500':
          description: Internal server error
  /quiz/{quizId}/comments:
    get:
      tags:
        - Quiz
      summary: Get all comments for a quiz
      security:
        - cookieAuth: []
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comments retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comments retrieved"
                  payload:
                    type: object
                    properties:
                      comments:
                        type: array
                        items:
                          type: object
        '500':
          description: Internal server error
...
  /admin/quiz/{id}/participants:
    get:
      tags:
        - Admin - Quiz
      summary: Search quiz participants (Teacher only)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Participants retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Participants retrieved successfully"
                  participants:
                    type: array
                    items:
                      type: object
                      properties:
                        username:
                          type: string
                        email:
                          type: string
        '400':
          description: Invalid quiz ID or not a live quiz
        '403':
          description: Not authorized
        '404':
          description: Quiz not found
        '500':
          description: Internal server error
...
  /admin/quiz/{quizId}/questions:
    post:
      tags:
        - Admin - Questions
      summary: Add question to quiz (Teacher only)
      security:
        - cookieAuth: []
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                questionText:
                  type: string
                questionType:
                  type: string
                  enum: [multiple-choice, true-false, short-answer]
                options:
                  type: array
                  items:
                    type: object
                correctAnswer:
                  type: string
                points:
                  type: number
                explanation:
                  type: string
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Question added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Question added successfully"
                  payload:
                    type: object
        '400':
          description: Validation error
        '403':
          description: Not authorized
        '404':
          description: Quiz not found
        '500':
          description: Internal server error
    get:
      tags:
        - Admin - Questions
      summary: Get all questions for quiz (Teacher only)
      security:
        - cookieAuth: []
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Questions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ok"
                  payload:
                    type: object
                    properties:
                      questions:
                        type: array
                        items:
                          type: object
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token